refactor: Folder structure improvements
