+++ examples/code-only/Example18_Box2DPhysics/Box2DPhysics/Core/PhysicsStepSettings.cs
@@ -1,4 +1,4 @@
-namespace Example18_Box2DPhysics.Reusable.Core;
+namespace Example18_Box2DPhysics.Box2DPhysics.Core;
 
 /// <summary>
 /// Configuration for fixed timestep simulation stepping. This is an extraction candidate from the
