+++ examples/code-only/Example18_Box2DPhysics/Box2DPhysics/Events/ContactEventType.cs
@@ -1,0 +1,8 @@
+namespace Example18_Box2DPhysics.Box2DPhysics.Events;
+
+public enum ContactEventType
+{
+    BeginTouch,
+    EndTouch,
+    Hit
+}
