+++ examples/code-only/Example18_Box2DPhysics/Box2DPhysics/Events/IContactEventHandler.cs
@@ -1,0 +1,6 @@
+namespace Example18_Box2DPhysics.Box2DPhysics.Events;
+
+public interface IContactEventHandler
+{
+    void OnContactEvent(ContactEventData eventData);
+}
