+++ examples/code-only/Example18_Box2DPhysics/Box2DPhysics/Events/ISensorEventHandler.cs
@@ -1,0 +1,6 @@
+namespace Example18_Box2DPhysics.Box2DPhysics.Events;
+
+public interface ISensorEventHandler
+{
+    void OnSensorEvent(SensorEventData eventData);
+}
