+++ examples/code-only/Example18_Box2DPhysics/Box2DPhysics/Events/SensorEventType.cs
@@ -1,0 +1,7 @@
+namespace Example18_Box2DPhysics.Box2DPhysics.Events;
+
+public enum SensorEventType
+{
+    BeginTouch,
+    EndTouch
+}
