+++ examples/code-only/Example18_Box2DPhysics/Physics/BodyForces.cs
@@ -11,20 +11,20 @@ public static class BodyForces
 {
     public static void ApplyImpulse(B2BodyId bodyId, Vector2 impulse)
     {
-        var b2Impulse = new Box2D.NET.B2Vec2(impulse.X, impulse.Y);
+        var b2Impulse = new B2Vec2(impulse.X, impulse.Y);
         b2Body_ApplyLinearImpulseToCenter(bodyId, b2Impulse, true);
     }
 
     public static void ApplyImpulseAtPoint(B2BodyId bodyId, Vector2 impulse, Vector2 point)
     {
-        var b2Impulse = new Box2D.NET.B2Vec2(impulse.X, impulse.Y);
-        var b2Point = new Box2D.NET.B2Vec2(point.X, point.Y);
+        var b2Impulse = new B2Vec2(impulse.X, impulse.Y);
+        var b2Point = new B2Vec2(point.X, point.Y);
         b2Body_ApplyLinearImpulse(bodyId, b2Impulse, b2Point, true);
     }
 
     public static void SetVelocity(B2BodyId bodyId, Vector2 velocity)
     {
-        var v = new Box2D.NET.B2Vec2(velocity.X, velocity.Y);
+        var v = new B2Vec2(velocity.X, velocity.Y);
         b2Body_SetLinearVelocity(bodyId, v);
     }
