+++ examples/code-only/Example18_Box2DPhysics/Program.cs
@@ -1,5 +1,6 @@
 using Example18_Box2DPhysics;
 using Example18_Box2DPhysics.Helpers;
+using Example18_Box2DPhysics.Box2DPhysics;
 using Stride.CommunityToolkit.Engine;
 using Stride.Core.Mathematics;
 using Stride.Engine;
